SQ3D-12 Add bounding boxes
A bounding box system is added to the AABB.h file
In Scene.cpp there is a HandleCollision function that handles the collisions between meshes.
In the Mesh.h file and the MeshActor actor class, the bounding interface is added which updates the bounding box center and extents to fit the mesh.

Actor controller added to just control any actor based on position.

UI is added to switch between controlling the player and camera

Fixed a memory leak when clearing cached resources.